+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Changes behaviour of `error_types` option when not set: before it defaulted to `error_reporting()` statically at SDK initialization; now it will be evaluated each time during error handling to allow silencing errors temporarily (#1196)
+- Fix enabling the `FatalErrorListenerIntegration`, `ErrorListenerIntegration` and `ExceptionListenerIntegration` integrations even when the DSN is not set (#1205)
 
 ## 3.2.0 (2021-03-03)